refactor: extract fixture functions to utils module
